combined results related redux files
