+++ packages/coinstac-ui/app/render/components/consortium/consortium-controller.js
@@ -16,9 +16,9 @@ import {
   saveConsortium,
 } from '../../state/ducks/consortia';
 import {
-  fetch as fetchRemoteResults,
+  fetchRemoteResults,
   setRemoteResults,
-} from '../../state/ducks/remote-results';
+} from '../../state/ducks/results';
 
 class ConsortiumController extends Component {
   constructor(props) {
@@ -171,7 +171,7 @@ function mapStateToProps(state, { params: { consortiumId, resultId } }) {
     computations,
     consortia,
     loading: { isLoading },
-    remoteResults,
+    resultsState: { remoteResults },
   } = state;
   const isNew = !consortiumId;
   const consortium = !isNew ?
